Fix wrong namespace reports -> report.
