+++ BlueTracker.SDK.Performance/Model/Processing/Report/Engine.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using BlueTracker.SDK.Performance.Model.Enums;
 
-namespace BlueTracker.SDK.Performance.Model.Processing.Reports
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
 {
     public class Engine
     {
