+++ BlueTracker.SDK.Performance/Model/Processing/Report/Environment.cs
@@ -1,6 +1,6 @@
 ﻿using BlueTracker.SDK.Performance.Model.Common;
 
-namespace BlueTracker.SDK.Performance.Model.Processing.Reports
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
 {
     public class Environment
     {
