+++ BlueTracker.SDK.Performance/Model/Processing/Report/FuelConsumer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace BlueTracker.SDK.Performance.Model.Processing.Reports
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
 {
     public abstract class FuelConsumer : Aggregate
     {
