+++ BlueTracker.SDK.Performance/Model/Processing/Report/FuelQuantity.cs
@@ -1,0 +1,20 @@
+﻿using BlueTracker.SDK.Performance.Model.Common;
+using BlueTracker.SDK.Performance.Model.Enums;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
+{
+    public class FuelQuantity
+    {
+        [JsonProperty(PropertyName = "kind")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public FuelKindOptions Kind { get; set; }
+
+        [JsonProperty(PropertyName = "amount")]
+        public double? Amount { get; set; }
+
+        [JsonProperty(PropertyName = "type")]
+        public FuelType Type { get; set; }
+    }
+}
