+++ BlueTracker.SDK.Performance/Model/Processing/Report/LubOilQuantity.cs
@@ -3,18 +3,20 @@ using BlueTracker.SDK.Performance.Model.Enums;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace BlueTracker.SDK.Performance.Model.Processing.Reports
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
 {
-    public class FuelQuantity
+    public class LubOilQuantity
     {
-        [JsonProperty(PropertyName = "kind")]
         [JsonConverter(typeof(StringEnumConverter))]
-        public FuelKindOptions Kind { get; set; }
+        public AggregateOptions Aggregate { get; set; }
+
+        [JsonConverter(typeof(StringEnumConverter))]
+        public LubOilKindOptions Kind { get; set; }
 
         [JsonProperty(PropertyName = "amount")]
         public double? Amount { get; set; }
 
         [JsonProperty(PropertyName = "type")]
-        public FuelType Type { get; set; }
+        public LubOilType Type { get; set; }
     }
 }
