+++ BlueTracker.SDK.Performance/Model/Processing/Report/PerformanceReport.cs
@@ -3,7 +3,7 @@ using BlueTracker.SDK.Performance.Model.Enums;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace BlueTracker.SDK.Performance.Model.Processing.Reports
+namespace BlueTracker.SDK.Performance.Model.Processing.Report
 {
     public class PerformanceReport
     {
